+++ internal/pkg/constants/tasks.go
@@ -2,4 +2,5 @@ package constants
 
 const TASK_ACTION_INSTALL = "install"
 const TASK_ACTION_DESTROY = "destroy"
+const TASK_ACTION_CLUSTER_CREATE = "cluster_create"
 const TASK_ACTION_CLUSTER_UPDATE = "cluster_update"